# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CPNS & School Tutoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List course packages
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a course package (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update package (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete package (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Bank Accounts
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, student sees own)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (buy a package)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created with payment instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/upload-proof:
    post:
      operationId: uploadPaymentProof
      tags: [orders]
      summary: Upload payment proof (student)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Proof uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/verify:
    post:
      operationId: verifyOrder
      tags: [orders]
      summary: Approve or reject an order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderVerifyInput"
      responses:
        "200":
          description: Order verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Enrollments
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions (admin/tutor)
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionList"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question (admin/tutor)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Quizzes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
...
    Package:
      type: object
      required: [id, name, description, price, durationDays, category, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        durationDays:
          type: integer
        category:
          type: string
          enum: [CPNS, SD, SMP, SMA]
        thumbnail:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        materialCount:
          type: ["integer", "null"]
        quizCount:
          type: ["integer", "null"]
        tryoutCount:
          type: ["integer", "null"]
    PackageDetail:
      type: object
      required: [id, name, description, price, durationDays, category, isActive, createdAt, materials, quizzes, tryouts]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        durationDays:
          type: integer
        category:
          type: string
        thumbnail:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        materials:
          type: array
          items:
            $ref: "#/components/schemas/Material"
        quizzes:
          type: array
          items:
            $ref: "#/components/schemas/Quiz"
        tryouts:
          type: array
          items:
            $ref: "#/components/schemas/Tryout"
    PackageInput:
      type: object
      required: [name, description, price, durationDays, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        durationDays:
          type: integer
        category:
          type: string
          enum: [CPNS, SD, SMP, SMA]
        thumbnail:
          type: ["string", "null"]
        isActive:
          type: boolean
    PackageUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        durationDays:
          type: integer
        category:
          type: string
        thumbnail:
          type: ["string", "null"]
        isActive:
          type: boolean

    # Bank Account
    BankAccount:
      type: object
      required: [id, bankName, accountNumber, accountHolder, isActive]
      properties:
        id:
          type: integer
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolder:
          type: string
        isActive:
          type: boolean
...
    Order:
      type: object
      required: [id, userId, packageId, orderCode, amount, uniqueAmount, status, expiredAt, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        packageId:
          type: integer
        orderCode:
          type: string
        amount:
          type: integer
        uniqueAmount:
          type: integer
        status:
          type: string
          enum: [PENDING, WAITING_VERIFICATION, PAID, EXPIRED, REJECTED]
        paymentProof:
          type: ["string", "null"]
        expiredAt:
          type: string
        paidAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
        package:
          $ref: "#/components/schemas/Package"
        bankAccounts:
          type: array
          items:
            $ref: "#/components/schemas/BankAccount"
    OrderInput:
      type: object
      required: [packageId]
      properties:
        packageId:
          type: integer
    OrderList:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PaymentProofInput:
      type: object
      required: [paymentProof]
      properties:
        paymentProof:
          type: string
    OrderVerifyInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject]
        rejectionReason:
          type: ["string", "null"]

    # Enrollment
...
    Material:
      type: object
      required: [id, packageId, title, orderIndex, createdAt]
      properties:
        id:
          type: integer
        packageId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        orderIndex:
          type: integer
        createdAt:
          type: string
        isRead:
          type: ["boolean", "null"]
...
    Question:
      type: object
      required: [id, questionText, optionA, optionB, optionC, optionD, optionE, correctAnswer, category, difficulty, createdAt]
      properties:
        id:
          type: integer
        questionText:
          type: string
        imageUrl:
          type: ["string", "null"]
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        optionE:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D, E]
        explanation:
          type: ["string", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
    QuestionInput:
      type: object
      required: [questionText, optionA, optionB, optionC, optionD, optionE, correctAnswer, category, difficulty]
      properties:
        questionText:
          type: string
        imageUrl:
          type: ["string", "null"]
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        optionE:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D, E]
        explanation:
          type: ["string", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: ["string", "null"]
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        imageUrl:
          type: ["string", "null"]
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        optionE:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D, E]
        explanation:
          type: ["string", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: ["string", "null"]
    QuestionList:
      type: object
      required: [questions, total, page, limit]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Quiz
    Quiz:
      type: object
      required: [id, packageId, title, timeLimit, passingScore, createdAt]
      properties:
        id:
          type: integer
        packageId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        timeLimit:
          type: integer
        passingScore:
          type: integer
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Tryout:
      type: object
      required: [id, title, type, durationMinutes, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [CPNS_SKD, UTBK, CUSTOM]
        durationMinutes:
          type: integer
        scheduledAt:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
...