# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPNS & School Tutoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionUpdateCorrectAnswer } from "./questionUpdateCorrectAnswer";
import type { QuestionUpdateDifficulty } from "./questionUpdateDifficulty";

export interface QuestionUpdate {
  questionText?: string;
  /** @nullable */
  imageUrl?: string | null;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  optionE?: string;
  correctAnswer?: QuestionUpdateCorrectAnswer;
  /** @nullable */
  explanation?: string | null;
  category?: string;
  difficulty?: QuestionUpdateDifficulty;
  /** @nullable */
  tags?: string | null;
}